Module google_analyticsdata1_beta::api

Structs

AnalyticsData
Central instance to access all AnalyticsData related resource activities

BatchRunPivotReportsRequest
The batch request containing multiple pivot report requests.

BatchRunPivotReportsResponse
The batch response containing multiple pivot reports.

BatchRunReportsRequest
The batch request containing multiple report requests.

BatchRunReportsResponse
The batch response containing multiple reports.

BetweenFilter
To express that the result needs to be between two numbers (inclusive).

CaseExpression
Used to convert a dimension value to a single case.

Cohort
Defines a cohort selection criteria. A cohort is a group of users who share a common characteristic. For example, users with the same firstSessionDate belong to the same cohort.

CohortReportSettings
Optional settings of a cohort report.

CohortSpec
The specification of cohorts for a cohort report. Cohort reports create a time series of user retention for the cohort. For example, you could select the cohort of users that were acquired in the first week of September and follow that cohort for the next six weeks. Selecting the users acquired in the first week of September cohort is specified in the cohort object. Following that cohort for the next six weeks is specified in the cohortsRange object. For examples, see Cohort Report Examples. The report response could show a weekly time series where say your app has retained 60% of this cohort after three weeks and 25% of this cohort after six weeks. These two percentages can be calculated by the metric cohortActiveUsers/cohortTotalUsers and will be separate rows in the report.

CohortsRange
Configures the extended reporting date range for a cohort report. Specifies an offset duration to follow the cohorts over.

ConcatenateExpression
Used to combine dimension values to a single dimension.

DateRange
A contiguous set of days: startDate, startDate + 1, ..., endDate. Requests are allowed up to 4 date ranges.

Dimension
Dimensions are attributes of your data. For example, the dimension city indicates the city from which an event originates. Dimension values in report responses are strings; for example, city could be "Paris" or "New York". Requests are allowed up to 8 dimensions.

DimensionExpression
Used to express a dimension which is the result of a formula of multiple dimensions. Example usages: 1) lower_case(dimension) 2) concatenate(dimension1, symbol, dimension2).

DimensionHeader
Describes a dimension column in the report. Dimensions requested in a report produce column entries within rows and DimensionHeaders. However, dimensions used exclusively within filters or expressions do not produce columns in a report; correspondingly, those dimensions do not produce headers.

DimensionMetadata
Explains a dimension.

DimensionOrderBy
Sorts by dimension values.

DimensionValue
The value of a dimension.

Filter
An expression to filter dimension or metric values.

FilterExpression
To express dimension or metric filters. The fields in the same FilterExpression need to be either all dimensions or all metrics.

FilterExpressionList
A list of filter expressions.

InListFilter
The result needs to be in a list of string values.

Metadata
The dimensions and metrics currently accepted in reporting methods.

Metric
The quantitative measurements of a report. For example, the metric eventCount is the total number of events. Requests are allowed up to 10 metrics.

MetricHeader
Describes a metric column in the report. Visible metrics requested in a report produce column entries within rows and MetricHeaders. However, metrics used exclusively within filters or expressions do not produce columns in a report; correspondingly, those metrics do not produce headers.

MetricMetadata
Explains a metric.

MetricOrderBy
Sorts by metric values.

MetricValue
The value of a metric.

NumericFilter
Filters for numeric or date values.

NumericValue
To represent a number.

OrderBy
The sort options.

Pivot
Describes the visible dimension columns and rows in the report response.

PivotDimensionHeader
Summarizes dimension values from a row for this pivot.

PivotHeader
Dimensions' values in a single pivot.

PivotOrderBy
Sorts by a pivot column group.

PivotSelection
A pair of dimension names and values. Rows with this dimension pivot pair are ordered by the metric's value. For example if pivots = {{"browser", "Chrome"}} and metric_name = "Sessions", then the rows will be sorted based on Sessions in Chrome. ---------|----------|----------------|----------|---------------- | Chrome | Chrome | Safari | Safari ---------|----------|----------------|----------|---------------- Country | Sessions | Pages/Sessions | Sessions | Pages/Sessions ---------|----------|----------------|----------|---------------- US | 2 | 2 | 3 | 1 ---------|----------|----------------|----------|---------------- Canada | 3 | 1 | 4 | 1 ---------|----------|----------------|----------|----------------

PropertyBatchRunPivotReportCall
Returns multiple pivot reports in a batch. All reports must be for the same GA4 Property.

PropertyBatchRunReportCall
Returns multiple reports in a batch. All reports must be for the same GA4 Property.

PropertyGetMetadataCall
Returns metadata for dimensions and metrics available in reporting methods. Used to explore the dimensions and metrics. In this method, a Google Analytics GA4 Property Identifier is specified in the request, and the metadata response includes Custom dimensions and metrics as well as Universal metadata. For example if a custom metric with parameter name levels_unlocked is registered to a property, the Metadata response will contain customEvent:levels_unlocked. Universal metadata are dimensions and metrics applicable to any property such as country and totalUsers.

PropertyMethods
A builder providing access to all methods supported on property resources. It is not used directly, but through the AnalyticsData hub.

PropertyQuota
Current state of all quotas for this Analytics Property. If any quota for a property is exhausted, all requests to that property will return Resource Exhausted errors.

PropertyRunPivotReportCall
Returns a customized pivot report of your Google Analytics event data. Pivot reports are more advanced and expressive formats than regular reports. In a pivot report, dimensions are only visible if they are included in a pivot. Multiple pivots can be specified to further dissect your data.

PropertyRunRealtimeReportCall
The Google Analytics Realtime API returns a customized report of realtime event data for your property. These reports show events and usage from the last 30 minutes.

PropertyRunReportCall
Returns a customized report of your Google Analytics event data. Reports contain statistics derived from data collected by the Google Analytics tracking code. The data returned from the API is as a table with columns for the requested dimensions and metrics. Metrics are individual measurements of user activity on your property, such as active users or event count. Dimensions break down metrics across some common criteria, such as country or event name.

QuotaStatus
Current state for a particular quota group.

ResponseMetaData
Response's metadata carrying additional information about the report content.

Row
Report data for each row. For example if RunReportRequest contains: none "dimensions": [ { "name": "eventName" }, { "name": "countryId" } ], "metrics": [ { "name": "eventCount" } ] One row with 'in_app_purchase' as the eventName, 'JP' as the countryId, and 15 as the eventCount, would be: none "dimensionValues": [ { "value": "in_app_purchase" }, { "value": "JP" } ], "metricValues": [ { "value": "15" } ]

RunPivotReportRequest
The request to generate a pivot report.

RunPivotReportResponse
The response pivot report table corresponding to a pivot request.

RunRealtimeReportRequest
The request to generate a realtime report.

RunRealtimeReportResponse
The response realtime report table corresponding to a request.

RunReportRequest
The request to generate a report.

RunReportResponse
The response report table corresponding to a request.

StringFilter
The filter for string

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.